saml Support

+++ powershell/ad_api/endpoints/DELETE/delete-website.ps1
@@ -14,6 +14,7 @@ param(
     
 )
 
+import-module webadministration
 # This Section Parses the body Parameter
 # You will need to customize this section to consume the Json correctly for your application
 $newbody = $body | ConvertFrom-Json
@@ -22,12 +23,19 @@ $site_name = $newbody.site_name
 #Write-Host "Sitename: $site_name"
 
 
-
+$websitebase = "c:\infra\websites"
   
- 
+ if(Test-Path IIS:/Sites/$site_name) {
+						
+		     Stop-WebSite -Name $site_name
+			 Remove-WebSite -Name $site_name 
+			 Remove-WebAppPool -Name $site_name 
+			
+			 Remove-Item $websitebase\$site_name -recurse -Force
+				
+ }
 
-Remove-WebSite -Name $site_name 
-Remove-WebAppPool -Name $site_name 
+
 
  #Sets the error message based on success or failure
